fix(workers): utilize vite ?worker imports.
