+++ vite.config.js
@@ -10,6 +10,9 @@ export default defineConfig(({ mode }) => {
 
     return {
         base: './',
+        worker: {
+            format: 'es',
+        },
         resolve: {
             alias: {
                 '!': '/node_modules',
